Adds APDUs for static Phonon functions. Still need to add tests
